Add mov reg. Opcode range 0x88..0x8b
Change hasDisp to DispSize. When dumping displacement, need to consider its
size. As displacement size is determined by ModR/M and SIB byte, I store
the information in Instruction.

For immediate value and memory offset, as they are associated with instruction,
I can encode the size information in operand type.