Parse OT_RELCB (code offset byte) operand.
To make jump instruction dump easier, I do not dump it's offset because
objdump calculates the offset to the begin of the nearest symbol.

Also I use the name jnz (name used in diStorm) instead of jne (used in
objdump) because I think jnz reflects it's depending on the Z flag more
directly.